test(autoware_traffic_light_arbiter): consolidate node tests into a single integration suite
Replace test_node.cpp and test_traffic_arbiter_characteristic.cpp with one
self-contained node-level integration suite covering ROS wiring (subscriptions,
parameter routing, publish, map gating, empty input). Exhaustive arbitration
logic stays covered ROS-free in test_traffic_light_arbiter_core.cpp (line
coverage maintained at ~95%). Drop the now-unused autoware_test_utils dependency.
