+++ perception/autoware_traffic_light_arbiter/package.xml
@@ -18,7 +18,6 @@
   <depend>autoware_lanelet2_utils</depend>
   <depend>autoware_map_msgs</depend>
   <depend>autoware_perception_msgs</depend>
-  <depend>autoware_test_utils</depend>
   <depend>lanelet2_core</depend>
   <depend>rclcpp</depend>
   <depend>rclcpp_components</depend>
